+++ CPAC/alff/alff.py
@@ -273,3 +273,38 @@ def alff_falff(wf, cfg, strat_pool, pipe_num, opt=None):
     }
 
     return (wf, outputs)
+
+def alff_falff_space_template(wf, cfg, strat_pool, pipe_num, opt=None):
+    '''
+    {"name": "alff_falff_space_template",
+     "config": ["amplitude_low_frequency_fluctuation"],
+     "switch": ["run"],
+     "option_key": "None",
+     "option_val": "None",
+     "inputs": [["space-template_desc-denoisedNofilt_bold"],
+                "space-template_desc-bold_mask"],
+     "outputs": ["space-template_alff",
+                 "space-template_falff"]}
+    '''
+
+    alff = create_alff(f'alff_falff_{pipe_num}')
+
+    alff.inputs.hp_input.hp = \
+        cfg.amplitude_low_frequency_fluctuation['highpass_cutoff']
+    alff.inputs.lp_input.lp = \
+        cfg.amplitude_low_frequency_fluctuation['lowpass_cutoff']
+    alff.get_node('hp_input').iterables = ('hp', alff.inputs.hp_input.hp)
+    alff.get_node('lp_input').iterables = ('lp', alff.inputs.lp_input.lp)
+
+    node, out = strat_pool.get_data(["space-template_desc-denoisedNofilt_bold"])
+    wf.connect(node, out, alff, 'inputspec.rest_res')
+
+    node, out = strat_pool.get_data("space-template_desc-bold_mask")
+    wf.connect(node, out, alff, 'inputspec.rest_mask')
+
+    outputs = {
+        'space-template_alff': (alff, 'outputspec.alff_img'),
+        'space-template_falff': (alff, 'outputspec.falff_img')
+    }
+
+    return (wf, outputs)