+++ CPAC/utils/configuration/__init__.py
@@ -16,9 +16,9 @@ License for more details.
 
 You should have received a copy of the GNU Lesser General Public
 License along with C-PAC. If not, see <https://www.gnu.org/licenses/>."""
-from .configuration import check_pname, Configuration, DEFAULT_PIPELINE_FILE, \
-                           Preconfiguration, set_subject
+from .configuration import check_pname, Configuration, Preconfiguration, \
+                           preconfig_yaml, set_subject
 from . import configuration, diff
 
-__all__ = ['check_pname', 'Configuration', 'configuration',
-           'DEFAULT_PIPELINE_FILE', 'diff', 'Preconfiguration', 'set_subject']
+__all__ = ['check_pname', 'Configuration', 'configuration', 'diff',
+           'Preconfiguration', 'preconfig_yaml', 'set_subject']